requirements.txt
blinker==1.7.0
click==8.1.7
confluent-kafka==2.3.0
Flask==3.0.2
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.0
#psycopg2==2.9.9
psycopg2-binary
python-dotenv
pyarrow==15.0.0
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
tzdata==2024.1
Werkzeug==3.0.1
